converted persons panel component to ts
