--- extension/src/PersonsPanel/constants/sort.js
@@ -1,9 +1,0 @@
-export const SORT_TYPE = {
-  alphabetically: "Alphabetically",
-  bookmarks: "Bookmarks",
-};
-
-export const SORT_ORDER = {
-  asc: "Ascending",
-  desc: "Descending",
-};
