+++ extension/src/PersonsPanel/constants/sort.ts
@@ -1,0 +1,9 @@
+export enum SORT_TYPE {
+  alphabetically = "Alphabetically",
+  bookmarks = "Bookmarks",
+}
+
+export enum SORT_ORDER {
+  asc = "Ascending",
+  desc = "Descending",
+}
