+++ extension/src/PersonsPanel/utils/sort.ts
@@ -1,6 +1,10 @@
 import { SORT_ORDER } from "../constants/sort";
+import { IPerson } from "../interfaces/persons";
 
-export const sortAlphabetically = (sortOrder, persons) => {
+export const sortAlphabetically = (
+  sortOrder: SORT_ORDER,
+  persons: IPerson[]
+) => {
   const sortedPersons = persons.sort((a, b) => a.name.localeCompare(b.name));
   if (sortOrder === SORT_ORDER.desc) {
     sortedPersons.reverse();
@@ -8,7 +12,10 @@ export const sortAlphabetically = (sortOrder, persons) => {
   return [...sortedPersons];
 };
 
-export const sortByBookmarksCount = (sortOrder, persons) => {
+export const sortByBookmarksCount = (
+  sortOrder: SORT_ORDER,
+  persons: IPerson[]
+) => {
   const sortedPersons = persons.sort(
     (a, b) => (a.taggedUrls?.length ?? 0) - (b.taggedUrls?.length ?? 0)
   );
