+++ extension/src/helpers/fetchFromStorage.ts
@@ -2,7 +2,10 @@ import storage from "ChromeApi/storage";
 import { EXTENSION_STATE, STORAGE_KEYS } from "GlobalConstants";
 import { UserInfo } from "SrcPath/HomePopup/interfaces/authentication";
 import { LastVisited } from "SrcPath/HomePopup/interfaces/lastVisited";
-import { Persons } from "SrcPath/PersonsPanel/interfaces/persons";
+import {
+  PersonImageUrls,
+  Persons,
+} from "SrcPath/PersonsPanel/interfaces/persons";
 import { Redirection } from "SrcPath/ShortcutsPanel/interfaces/redirections";
 
 export const getExtensionState = async () => {
@@ -43,4 +46,11 @@ export const getPersons = async () => {
     STORAGE_KEYS.persons
   );
   return persons as Persons;
+};
+
+export const getPersonImageUrls = async () => {
+  const { [STORAGE_KEYS.personImageUrls]: personImageUrls } = await storage.get(
+    STORAGE_KEYS.personImageUrls
+  );
+  return personImageUrls as PersonImageUrls;
 };
