+++ extension/src/utils/bypass/index.js
@@ -47,8 +47,8 @@ export const bypass = async (tabId, url) => {
 };
 
 export const syncBypassToStorage = async () => {
-  const snapshot = await getFromFirebase(FIREBASE_DB_REF.bypass);
-  const bypass = getMappedBypass(snapshot.val());
+  const response = await getFromFirebase(FIREBASE_DB_REF.bypass);
+  const bypass = getMappedBypass(response);
   await storage.set({ [STORAGE_KEYS.bypass]: bypass });
   console.log("Bypass is set to", bypass);
 };
